Refactor: 역할 리스트 공통화
